Split run.c into multiple files each relating to their function
