+++ lib/ak_softmax.h
@@ -1,0 +1,28 @@
+#ifndef _ak_softmax_H
+#define _ak_softmax_H
+
+//  Softmax exponentiates its input and normalizes it to create a probability distribution
+//  where the sum of the elements is 1.  This alters the array x inline.
+
+static inline
+void ak_softmax(float* x, int size) {
+    // find max value (for numerical stability)
+    float max_val = x[0];
+    for (int i = 1; i < size; i++) {
+        if (x[i] > max_val) {
+            max_val = x[i];
+        }
+    }
+    // exp and sum
+    float sum = 0.0f;
+    for (int i = 0; i < size; i++) {
+        x[i] = expf(x[i] - max_val);
+        sum += x[i];
+    }
+    // normalize
+    for (int i = 0; i < size; i++) {
+        x[i] /= sum;
+    }
+}
+
+#endif
